// src/app/pages/customers/interfaces/customer-response.interface.ts
// Generated by https://quicktype.io

export interface CustomerResponse {
    ok:       boolean;
    customers?: Customer[];
    customer?: Customer;
}

export interface Customer {
    _id?:            string;
    documentNumber: string;
    name:           string;
    address?:        string;
    phoneNumber:    string;
    createdAt?:      string;
    updatedAt?:      string;
}